Update dependencies and and update to lavalink v4, shoukaku v4

+++ src/events/client/SetupSystem.ts
@@ -15,7 +15,7 @@ export default class SetupSystem extends Event {
         if (!(channel instanceof TextChannel)) return;
         if (!message.member.voice.channel) {
             await oops(channel, `You are not connected to a voice channel to queue songs.`);
-            if (message) await message.delete().catch(() => {});
+            if (message) await message.delete().catch(() => { });
             return;
         }
 
@@ -28,22 +28,21 @@ export default class SetupSystem extends Event {
                 channel,
                 `I don't have enough permission to connect/speak in <#${message.member.voice.channel.id}>`
             );
-            if (message) await message.delete().catch(() => {});
+            if (message) await message.delete().catch(() => { });
             return;
         }
 
         if (
             message.guild.members.cache.get(this.client.user.id).voice.channel &&
             message.guild.members.cache.get(this.client.user.id).voice.channelId !==
-                message.member.voice.channelId
+            message.member.voice.channelId
         ) {
             await oops(
                 channel,
-                `You are not connected to <#${
-                    message.guild.members.cache.get(this.client.user.id).voice.channelId
+                `You are not connected to <#${message.guild.members.cache.get(this.client.user.id).voice.channelId
                 }> to queue songs`
             );
-            if (message) await message.delete().catch(() => {});
+            if (message) await message.delete().catch(() => { });
             return;
         }
         let player = this.client.queue.get(message.guildId);
@@ -52,11 +51,11 @@ export default class SetupSystem extends Event {
                 message.guild,
                 message.member.voice.channel,
                 message.channel,
-                this.client.shoukaku.getNode()
+                this.client.shoukaku.options.nodeResolver(this.client.shoukaku.nodes)
             );
         }
         await setupStart(this.client, message.content, player, message);
-        if (message) await message.delete().catch(() => {});
+        if (message) await message.delete().catch(() => { });
     }
 }
